Remove gosu in favour of alpines iputils-ping
Not only is the iputils-ping package smaller, it correctly works
within a docker container.

The side effect of removing gosu is that docker containers built
without it show ~52 fewer security vulnerabilities without the
gosu binary installed.

While the author of gosu says this isn't a problem anyway, its
just a bad practice. As such, we end up with a smaller window
for issues, and a smaller container size as a bonus.